+++ R/penhdfeppml_int.R
@@ -26,6 +26,14 @@
 #'     coefficient-specific penalty weights (see details). Otherwise, a single global penalty is used.
 #' @param debug Logical. If \code{TRUE}, this helps with debugging penalty weights by printing output
 #'    of the first iteration to the console and stopping the estimation algorithm.
+#' @param colcheck_x Logical. If \code{TRUE}, this checks collinearity between the independent variables and drops the
+#' collinear variables.
+#' @param colcheck_x_fes Logical. If \code{TRUE}, this checks whether the independent variables are perfectly explained
+#' by the fixed effects drops those that are perfectly explained.
+#' @param gamma_val Numerical value that determines the regularization threshold as defined in Belloni, Chernozhukov, Hansen, and Kozbur (2016). NULL default sets parameter to 0.1/log(n).
+#' @param phipost Logical. If \code{TRUE}, the plugin coefficient-specific penalty weights are iteratively
+#' calculated using estimates from a post-penalty regression when \code{method == "plugin"}. Otherwise,
+#' these are calculated using estimates from a penalty regression.
 #' @inheritParams hdfeppml_int
 #'
 #' @return If \code{method == "lasso"} (the default), an object of class \code{elnet} with the elements
@@ -66,11 +74,16 @@
 #' \donttest{reg <- penhdfeppml_int(y = y, x = x, fes = fes, lambda = 0.1, penalty = "ridge")}
 #'
 #' @inheritSection hdfeppml_int References
+#' @importFrom stats gaussian
+#' @importFrom stats var
+#' @importFrom utils head
+#' @importFrom devtools load_all
 
 penhdfeppml_int <- function(y, x, fes, lambda, tol = 1e-8, hdfetol = 1e-4, glmnettol = 1e-12,
-                            penalty = "lasso", penweights = NULL, saveX = TRUE, mu = NULL, colcheck_x = TRUE, colcheck_x_fes = TRUE,
-                            init_z = NULL, post = FALSE, verbose = FALSE, standardize = TRUE,
-                            method = "placeholder", cluster = NULL, debug = FALSE) {
+                            penalty = "lasso", penweights = NULL, saveX = TRUE, mu = NULL, colcheck_x = FALSE, colcheck_x_fes = TRUE,
+                            init_z = NULL, post = FALSE, verbose = FALSE, phipost = TRUE, standardize = TRUE,
+                            method = "placeholder", cluster = NULL, debug = FALSE, gamma_val=NULL) {
+  xnames <- colnames(x)
   old_x <- x
   old_y <- y
   old_fes <- fes
@@ -80,7 +93,7 @@ penhdfeppml_int <- function(y, x, fes, lambda, tol = 1e-8, hdfetol = 1e-4, glmne
                                       hdfetol = hdfetol, glmnettol = glmnettol, penalty = penalty,
                                       penweights = penweights, saveX = saveX,
                                       mu = mu, colcheck_x = colcheck_x, colcheck_x_fes = colcheck_x_fes, K = 15, init_z = init_z, post = FALSE,
-                                      verbose = verbose, lambda = NULL)
+                                      verbose = verbose, phipost=phipost, lambda = NULL, gamma_val=gamma_val)
   }
   else {
 
@@ -89,9 +102,9 @@ penhdfeppml_int <- function(y, x, fes, lambda, tol = 1e-8, hdfetol = 1e-4, glmne
     rownames(b) <- colnames(x)
     include_x <- 1:ncol(x)
 
-    if (is.null(penweights)) {
-      penweights <- rep(1, length(include_x))   #note this is the default used by glmnet. Using "NULL" actually gives you incorrect weights.
-    }
+   # if (is.null(penweights)) {
+   #    penweights <- rep(1, length(include_x))   #note this is the default used by glmnet. Using "NULL" actually gives you incorrect weights.
+   #  } #N: moved this below colchecks, because include_x gets updated
 
     if (length(penweights) > length(include_x)){
       print("penweights needs to be same length as number of x variables")
@@ -99,17 +112,37 @@ penhdfeppml_int <- function(y, x, fes, lambda, tol = 1e-8, hdfetol = 1e-4, glmne
     }
 
     # collinearity check
-    if (colcheck_x == TRUE | colcheck_x_fes == TRUE) {
-      include_x <- collinearity_check(y, x, fes, 1e-6, colcheck_x = colcheck_x, colcheck_x_fes = colcheck_x_fes)
-      x <- x[, include_x]
-      if (!is.null(penweights)) {
-        penweights = penweights[include_x]
-      }
-      colcheck_x_post <- FALSE
-      colcheck_x_fes_post <- FALSE
-    } else {
-      colcheck_x_post <- TRUE
-      colcheck_x_fes_post <- TRUE
+    if (colcheck_x==TRUE & colcheck_x_fes==TRUE){
+      include_x <- collinearity_check(y,x,fes,1e-6, colcheck_x=colcheck_x, colcheck_x_fes=colcheck_x_fes)
+      x <- x[,include_x]
+      colnames(x) <- xnames[include_x]
+      xnames <- xnames[include_x]
+      colcheck_x_post = FALSE
+      colcheck_x_fes_post = FALSE
+    }
+    if (colcheck_x==FALSE & colcheck_x_fes==FALSE){
+      colcheck_x_post = TRUE
+      colcheck_x_fes_post = TRUE
+    }
+    if (colcheck_x==TRUE & colcheck_x_fes==FALSE){
+      include_x <- collinearity_check(y,x,fes,1e-6, colcheck_x=colcheck_x, colcheck_x_fes=colcheck_x_fes)
+      x <- x[,include_x]
+      colnames(x) <- xnames[include_x]
+      xnames <- xnames[include_x]
+      colcheck_x_post = TRUE
+      colcheck_x_fes_post = FALSE
+    }
+    if (colcheck_x==FALSE & colcheck_x_fes == TRUE){
+      include_x <- collinearity_check(y,x,fes,1e-6, colcheck_x=colcheck_x, colcheck_x_fes=colcheck_x_fes)
+      x <- x[,include_x]
+      colnames(x) <- xnames[include_x]
+      xnames <- xnames[include_x]
+      colcheck_x_post = FALSE
+      colcheck_x_fes_post = TRUE
+    }
+
+    if (is.null(penweights)) {
+      penweights <- rep(1, length(include_x))   #note this is the default used by glmnet. Using "NULL" actually gives you incorrect weights.
     }
 
     # number of obs (needed for deviance)
@@ -119,7 +152,6 @@ penhdfeppml_int <- function(y, x, fes, lambda, tol = 1e-8, hdfetol = 1e-4, glmne
     old_deviance <-0
     iter <-0
     while (crit > tol) {
-      #print(iter)
       iter <- iter + 1
 
       if(iter > 50){message("Lasso exceeded 50 iterations. Break loop and return last model."); break}
@@ -131,7 +163,7 @@ penhdfeppml_int <- function(y, x, fes, lambda, tol = 1e-8, hdfetol = 1e-4, glmne
                                    init_z = NULL, verbose = FALSE, maxiter = 1000, cluster = NULL, vcv = TRUE)
           mu <- only_fes$mu
           #      mu <- mu[mu>0]
-          #  print(length(mu))
+          # print(length(mu))
         }
         z   <- (y - mu)/mu + log(mu)
         #     z <- z[mu>0]
@@ -192,22 +224,22 @@ penhdfeppml_int <- function(y, x, fes, lambda, tol = 1e-8, hdfetol = 1e-4, glmne
         }
       }
 
-      if(sd(z_resid)==Inf){
-        print("z_resid")
-        print(head(z_resid))
-        print(sum(is.na(z_resid)))
-        print(mean(z_resid))
-        print(sd(z_resid))
-      }
-      if(length(reg_z)!=length(z_resid)){
-        mu_t <- mu
-        reg_z_t <- reg_z
-        temp1 <- (collapse::fhdwithin(reg_z, fes, w = mu, eps = hdfetol))
-        temp2 <- (lfe::demeanlist(reg_z, fes, weights = sqrt(mu), eps = hdfetol))
-        print("check")
-        print(all.equal(temp1,temp2))
-        print(sum(is.na(temp2)))
-      }
+      # if(sd(z_resid)==Inf){
+      #   print("z_resid")
+      #   print(head(z_resid))
+      #   print(sum(is.na(z_resid)))
+      #   print(mean(z_resid))
+      #   print(sd(z_resid))
+      # }
+      # if(length(reg_z)!=length(z_resid)){
+      #   mu_t <- mu
+      #   reg_z_t <- reg_z
+      #   temp1 <- (collapse::fhdwithin(reg_z, fes, w = mu, eps = hdfetol))
+      #   temp2 <- (lfe::demeanlist(reg_z, fes, weights = sqrt(mu), eps = hdfetol))
+      #   print("check")
+      #   print(all.equal(temp1,temp2))
+      #   print(sum(is.na(temp2)))
+      # }
 
       #       print(length(z))
       #       print(length(reg_z))
@@ -222,13 +254,16 @@ penhdfeppml_int <- function(y, x, fes, lambda, tol = 1e-8, hdfetol = 1e-4, glmne
       # print(dim(x_resid))
 
       if (is.null(penweights)) {
+        print("null penweights")
         #penreg <- glmnet(x = x_resid, y = z_resid, weights = mu/sum(mu), lambda = lambda, thresh = glmnettol, standardize = standardize, family=gaussian(link = "identity"), warm.g=NULL)
       }
       else{
         if(debug) {
+          print("sum penweights")
           print(sum(penweights))
         }
         if(debug) {
+          print("penweights")
           print(penweights)
         }
         penweights = penweights * length(include_x) / sum(penweights)
@@ -254,18 +289,18 @@ penhdfeppml_int <- function(y, x, fes, lambda, tol = 1e-8, hdfetol = 1e-4, glmne
           warning(penalty, " penalty is not supported. Lasso is used by default.")
         }
         if (debug) {
-          penreg <- glmnet(x = x_resid, y = z_resid, weights = mu/sum(mu), lambda = lambda,
+          penreg <- glmnet::glmnet(x = x_resid, y = z_resid, weights = mu/sum(mu), lambda = lambda,
                                 thresh = glmnettol, standardize = standardize, family=gaussian(link = "identity"))
           print((penreg$beta))
           print(penweights)
         }
         if (debug) {
-          penreg <- glmnet(x = x_resid, y = z_resid, weights = mu/sum(mu), lambda = lambda,
+          penreg <- glmnet::glmnet(x = x_resid, y = z_resid, weights = mu/sum(mu), lambda = lambda,
                                 thresh = glmnettol, standardize = standardize, family=gaussian(link = "identity"))
           print((penreg$beta))
           print(penweights)
         }
-        penreg <- glmnet(x = x_resid, y = z_resid, weights = mu/sum(mu), lambda = lambda,
+        penreg <- glmnet::glmnet(x = x_resid, y = z_resid, weights = mu/sum(mu), lambda = lambda,
                               thresh = glmnettol, penalty.factor = penweights, standardize = standardize, family=gaussian(link = "identity"))
         if (debug) {
           print((penreg$beta))