+++ NEWS.md
@@ -1,3 +1,14 @@
+# penppml 0.2.1.900
+
+* When post = FALSE, there was a mistake in the last version where a non-existent
+* vector of coefficients was being stored. This issue is now corrected.
+
+# penppml 0.2.0.900
+
+* bootstrap function added
+* gamma_val option added for plugin Lasso
+* Two possible collinearity checks introduced
+
 # penppml 0.1.1.900
 
 * When standardize option = TRUE and selected penalty is ridge, there was a mistake in last